+++ Assignment1/Assignment1/Platform.cpp
@@ -30,8 +30,8 @@ Platform::~Platform() {
 void Platform::CreateCollision(int i) {
 	mCollision.x = mPlatform[i]->Position().x;
 	mCollision.y = mPlatform[i]->Position().y - (mPlatform[i]->GetHeight() / 2) + 5;
-	mCollision.w = mPlatform[i]->GetWidth();
-	mCollision.h = 10;
+	mCollision.w = (float)mPlatform[i]->GetWidth();
+	mCollision.h = 10.0f;
 }
 
 void Platform::Render(int i) {
